LayoutTests/platform/chromium/inspector/runtime/runtime-getProperties-expected.txt

Tests RemoteObject.getProperties.


Running: testSetUp

Running: testGetterAndSetter
{
    name : "get foo"
    value : {
        type : "function"
        description : "function foo() { return 1; }"
    }
}
{
    name : "set foo"
    value : {
        type : "function"
        description : "function foo(value) { }"
    }
}

Running: testGetterOnly
{
    name : "get foo"
    value : {
        type : "function"
        description : "function foo() { return 1; }"
    }
}
